+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Minor change to the `Notification`'s heading component so it adds a proper id to `NotificationHeading`.
 - Updates the `Heading` component to use the DS `Link` component rather than Chakra's `Link` component.
 - Updates the style of links rendered in the `Heading` component so they are styled as standard links. The only exception is for links used in the `Card` component for the _full-click functionality_ feature. Otherwise, links in `Card` headings are styled as standard links.
+- Updates the `SearchBar` component to now be implemented with the `ComponentWrapper` component.
 - Removes the `ButtonTypes.SearchBar` variant style for the `Button` component. The style object is now set and passed directly to the `Button` component in the `SearchBar` component through the `additionalStyles` prop.
 
 ### Fixes
@@ -52,7 +53,6 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the spacing in the `Notification` component to improve the alignment of the icon and text elements.
 - Updates the `Notification` component to optionally display/hide the `Icon` in the heading or content area.
 - Updates the `ComponentWrapper` component to allow the internal `HelperErrorText` component to get set in the invalid state. Also added tests for this component.
-- Updates the `SearchBar` component to now be implemented with the `ComponentWrapper` component.
 
 ### Fixes